requirements.txt
streamlit==1.55.0

google-genai==1.68.0
python-dotenv==1.2.2

langchain-pinecone==0.2.13
pinecone>=6.0.0

langchain==1.2.13
langchain-google-genai==4.2.1

pypdf==6.9.1

langchain-text-splitters==1.1.1

langchain-community>=0.3.31

langchain-classic>=1.0.1